+++ mango3-accounts/src/pages/mod.rs
@@ -1,5 +1,7 @@
 mod login_page;
 mod register_page;
+mod reset_password_page;
 
 pub use login_page::LoginPage;
 pub use register_page::RegisterPage;
+pub use reset_password_page::ResetPasswordPage;