+++ mango3-my-account/src/pages/index_parent_page.rs
@@ -30,7 +30,7 @@ pub fn IndexParentPage() -> impl IntoView {
             ("/", t_string!(i18n, shared.home)),
             ("/edit-profile", t_string!(i18n, my_account.edit_profile)),
             ("/edit-email", t_string!(i18n, my_account.edit_email)),
-            ("/change-password", t_string!(i18n, my_account.change_password)),
+            ("/change-password", t_string!(i18n, shared.change_password)),
         ]
     };